requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
apipkg==1.5
Appium-Python-Client==1.0.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
execnet==1.7.1
idna==2.10
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.0
pytest-forked==1.3.0
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.0
PyYAML==5.3.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
